[i26] Finished implementation of pizza cube use
I had the pizza events (cooking and gaining effects) as members of the
turns in which they occurred.  The display worked perfectly.  It seems
to me that everything should be done this way, and not going through
single lists spanning the entire quest, stopping at a particular turn
number, and hoping the turn chosen with that number is the right one.  I
should look into changing this, for accuracy.